+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.9.6 -
 
 - Extended the Keyboard shortcuts for the main practice page in the app and also added visual indicators. Fixes [#38](https://github.com/Razen04/GateQuest/issues/38).
+- Numerical Answers now support range based answers, exact answers, multiple answers and answers with tolerance as these happen very frequently in competitive exams. Fixes [#39](https://github.com/Razen04/GateQuest/issues/39).
 
 # v0.9.5 - 2026-02-08